refactor: 회원가입시 플랫폼 지정
- 사용하지 않은 생성자 제거
- 플랫폼 매직 넘버로 선언
- 회원 생성시 프인트 0, 사용하는 플랫폼으로 객체 생성
